FAC-683: Refactor - Repackaging UseCase from `question` Package to `questionimpact` package
