+++ flickit-assessment-kit/src/main/java/org/flickit/assessment/kit/adapter/in/rest/questionimpact/GetQuestionImpactsRestController.java
@@ -1,9 +1,9 @@
-package org.flickit.assessment.kit.adapter.in.rest.question;
+package org.flickit.assessment.kit.adapter.in.rest.questionimpact;
 
 import lombok.RequiredArgsConstructor;
 import org.flickit.assessment.common.config.jwt.UserContext;
-import org.flickit.assessment.kit.application.port.in.question.GetQuestionImpactsUseCase;
-import org.flickit.assessment.kit.application.port.in.question.GetQuestionImpactsUseCase.*;
+import org.flickit.assessment.kit.application.port.in.questionimpact.GetQuestionImpactsUseCase;
+import org.flickit.assessment.kit.application.port.in.questionimpact.GetQuestionImpactsUseCase.*;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
